chore: :wrench: Add CI/CD pipeline and deployment script
This commit adds a CI/CD pipeline for deploying the application and a deployment script for copying project files to an EC2 instance and starting the application. It also includes commands to stop and remove containers related to the 'crafty-cotton-api' image, prune unused images and volumes, and start the application with the necessary environment variables and dependencies.